fix: apply suggested changes after review
